move functions to widgets
